pyproject.toml
[tool.poetry]
name = "mariaai-authenticator"
version = "0.1.0"
description = "A Python project leveraging SyaNet to authenticate images with advanced AI models."
authors = ["Your Name <israel.santosmonteiro@gmail.com>"]
license = "MIT"
readme = "docs/README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pillow = "^11.1.0"
qrcode = "^8.0"

[tool.poetry.dev-dependencies]
pytest = ">=8.3.4,<9.0.0"
black = "^23.9.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

